phase-2/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.16
psycopg2-binary==2.9.9
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.1.0

python-multipart>=0.0.6
